# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kimyo v2 API Documentation",
        "description": "9-sinf o'quvchilari uchun interaktiv kimyo platformasi API hujjatlari",
        "contact": {
            "email": "admin@kimyo.uz"
        },
        "version": "2.5.13"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Asosiy API Server"
        }
    ],
    "paths": {
...
        "/api/v1/mobile/reference/schools": {
            "get": {
                "tags": [
                    "Mobile Reference"
                ],
                "summary": "Faol maktablar ro‘yxati",
                "operationId": "e9ee7f41e7ba83de4dd406fc975d2a54",
                "responses": {
                    "200": {
                        "description": "JSend success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mobile Reference"
                ],
                "summary": "Yangi maktab (faqat admin)",
                "operationId": "2ea9202bbc8aee9734bc34610ca786e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "15-sonli maktab"
                                    },
                                    "short_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "region": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yaratildi"
                    },
                    "403": {
                        "description": "Admin emas"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/mobile/reference/grades": {
            "get": {
                "tags": [
                    "Mobile Reference"
                ],
                "summary": "Sinf (grade) qiymatlari",
                "operationId": "3458c01af95f24faa48434e2d0b1fc9f",
                "responses": {
                    "200": {
                        "description": "JSend success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mobile Reference"
                ],
                "summary": "Yangi sinf qiymati (faqat admin)",
                "operationId": "82afee35163848bc5954d6e4ea93179a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "9-A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yaratildi"
                    },
                    "403": {
                        "description": "Admin emas"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum Bearer Token orqali autentifikatsiya",
                "name": "Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "MobileApiKey": {
                "type": "apiKey",
                "description": "Barcha so'rovlarda majburiy. Header: X-API-Key: <key>",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}